fix more warnings

+++ source/adios2/engine/ssc/SscReaderGeneric.cpp
@@ -576,7 +576,7 @@ SscReaderGeneric::BlocksInfo(const VariableStruct &variable,
 {
     std::vector<VariableStruct::BPInfo> ret;
     size_t blockID = 0;
-    for (int i = 0; i < m_GlobalWritePattern.size(); ++i)
+    for (int i = 0; i < static_cast<int>(m_GlobalWritePattern.size()); ++i)
     {
         for (auto &v : m_GlobalWritePattern[i])
         {
